+++ app/src/main/java/com/live/emmazone/model/DeliveryTemplateModel.kt
@@ -7,11 +7,11 @@ data class DeliveryTemplateModel(
     @SerializedName("id")
     var id: Int = 0,
     @SerializedName("shop_price")
-    var shop_price: Int = 0,
+    var shop_price: Double = 0.00,
     @SerializedName("bicycle_price")
-    var bicycle_price: Int = 0,
+    var bicycle_price: Double = 0.00,
     @SerializedName("logistics_price")
-    var logistics_price: Int = 0,
+    var logistics_price: Double = 0.00,
     @SerializedName("vendorId")
     var vendorId: Int = 0,
     @SerializedName("templateNo")
